Run the model for any variable

Run a linear model of the variable vs mpg for the mtcars data set and extract the r squared.

Usage

do_mod(x)

Arguments

x

variable name from mtcars

Value

data.frame with the variable name and r.squared value